Revision del codigo ya que adnaba en PostMan, pero enno funcionaba con el Front

+++ BackEnd-SpringBoot/src/main/java/com/portfolio/backend/security/repository/RoleRepository.java
@@ -1,12 +1,12 @@
 package com.portfolio.backend.security.repository;
 
-import com.portfolio.backend.security.models.Role;
 import com.portfolio.backend.security.enums.RoleName;
+import com.portfolio.backend.security.model.Role;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long>{
+public interface RoleRepository extends JpaRepository<Role, Integer>{
     Optional<Role> findByRoleName(RoleName roleName);
 }
